added toast for network error; renamed files and directory of ErrorPage: 'not-found' -> 'error'
