+++ src/components/pages/index.ts
@@ -1,7 +1,7 @@
 import FavoritesPage from './favorites/favorites.tsx';
 import LoginPage from './login/login.tsx';
 import MainPage from './main/main.tsx';
-import ErrorPage from './not-found/not-found.tsx';
+import ErrorPage from './error/error.tsx';
 import OfferPage from './offer/offer.tsx';
 
 export {
